task2/requirements.txt

mysql-connector-python==8.0.28
Faker==13.3.0
python-dotenv==0.19.2
